refactor(main) create new package config
